requirements.txt
openai>=1.0.0
faiss-cpu
streamlit
streamlit-chat
numpy
python-dotenv
requests
beautifulsoup4